fix box-sizing of <input>, <textarea>, <select>, <fieldset> + minor fixes
The <input>, <textarea>, <select> and <fieldset> html elements in
index.html have an anomalous behaviour and appear larger than the
parent <form> element, and the width/height are apparently partially
ignored. This problem did not show up in the FreeCodeCamp in-built
online code editor with live preview.

This is due to the CSS property 'box-sizing'. Setting this property
to 'box-sizing: border-box;' fixes the problem.

Indeed, from https://www.w3schools.com/csS/css3_box-sizing.asp, by default,
the actual width and height of an element are calculated as:

   actual width of an element  = width + padding + border
   actual height of an element = height + padding + border

and when setting the width/height of an element, this often appears bigger,
as its border and padding are added to the the specified width/height.
This is like setting 'box-sizing: content-box;'.

This problem is solved with the 'box-sizing: border-box;' property, as now

   actual width of an element  = width  = content width  + padding + border
   actual height of an element = height = content height + padding + border

Some browser already use box-sizing: border-box; for many form elements (but
not all - which is why inputs and text areas look different at width: 100%;).
Possibly, this is why the problem did not occur using FreeCodeCamp live preview.

Since the result of using the 'box-sizing: border-box;' is so much better,
many developers want all elements on their pages to work this way. Then,
applying this to all elements is safe and wise:

   * {
     box-sizing: border-box;
   }

Some minor adjustments to the padding of the <input>, <textarea>, <select> and
the max/min width of the <form> elements have also been applied.

Also, a CSS normalizer (https://necolas.github.io/normalize.css/8.0.1/normalize.css)
has been included, to possibly avoid different behaviours in different browsers.

Finally, a typo has been corrected in a label in the form in index.html, from
'Enter your name Email' to 'Enter your name email'.